High-Level API

Both the low- and high-level APIs start off with an instance of SemanticScholarConnection, which handles managing the request rate.

julia> using SemanticScholar

julia> s2c = SemanticScholarConnection()
SemanticScholarConnection(Dates.DateTime[])

This s2c struct is then passed to all of the functions to make requests to the Semantic Scholar API.

Paper Related

SemanticScholar.search_papers — Function

search_papers(s2c, query; limit=100)

Search the S2 Academic Graph for papers matching the search query.

This is a high-level API function, so it returns an array of Papers.

Author Related

SemanticScholar.search_authors — Function

search_authors(s2c, query; limit=50)

Search the S2 Academic Graph for authors with names matching the search query.

This is a high-level API function, so it returns an array of Authors.